+++ src/consts/consts.js
@@ -1,4 +1,4 @@
-import {getRandomInteger} from "../utils.js";
+import {getRandomInteger} from "../utils/common.js";
 
 export const POINT_COUNT = 20;
