+++ src/mock/route.js
@@ -1,4 +1,4 @@
-import {getRandomInteger} from "../utils.js";
+import {getRandomInteger} from "../utils/common.js";
 import dayjs from 'dayjs';
 
 export const generateRouteInfo = (points) => {
