+++ src/utils/render.js
@@ -1,0 +1,31 @@
+import Abstract from "../view/abstract/abstract.js";
+
+export const renderTemplate = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`
+};
+
+export const renderElement = (container, child, place) => {
+  if (container instanceof Abstract) {
+    container = container.getElement();
+  }
+  if (child instanceof Abstract) {
+    child = child.getElement();
+  }
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(child);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(child);
+      break;
+    case RenderPosition.AFTEREND:
+      container.after(child);
+      break;
+  }
+};
